psmouse: Skip psmouse_initialize for FocalTech touchpads.
The driver used to skip all initialization code for FocalTech touchpads
because they stop responding when certain commands are sent. My previous
patch failed to achieve the same and psmouse_reconnect would still cause
standard initialization code to be executed.

This patch introduces a field "skip_standard_init" to struct psmouse. If
this field is set to true, psmouse_initialize will not do anything. This
solution is somewhat ugly, but I believe it is the least ugly way to
introduce this apparently required special case.

Also, the code is modified to not send PSMOUSE_CMD_ENABLE twice (once in
focaltech.c, once in the generic code in psmouse-base.c).